requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
aiosqlite>=0.20
beautifulsoup4>=4.12
jinja2>=3.1
python-multipart>=0.0.9
pyyaml>=6.0

